3-feat-feedback-view: AddFeedbackNavigationItem에서 XButton 분리
